move spmemvfs
