+++ gframe/data_manager.h
@@ -3,7 +3,7 @@
 
 #include "config.h"
 #include "sqlite3.h"
-#include "spmemvfs.h"
+#include "spmemvfs/spmemvfs.h"
 #include "client_card.h"
 #include <unordered_map>
