+++ gframe/spmemvfs/premake4.lua
@@ -1,0 +1,6 @@
+project "cspmemvfs"
+    kind "StaticLib"
+    files { "**.c", "**.h" }
+
+    configuration "windows"
+        includedirs { "../../sqlite3" }
