+++ gframe/spmemvfs/spmemvfs.c
