+++ gframe/spmemvfs/spmemvfs.h
